+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - add werror, info, debug options.
 - modify parse front matter logic.
+- add for each values of object in filters.
 
 ## [0.4.0] - 2022-04-09