+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 
+## [0.5.0] - 2022-04-30
+
 - add werror, info, debug options.
 - modify parse front matter logic.
 - add for each values of object in filters.
@@ -37,7 +39,8 @@
 
 - initial release (build, server, info, list, output, md, njk, relurl).
 
-[Unreleased]: https://github.com/ankys/homura_deno/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/ankys/homura_deno/compare/v0.5.0...HEAD
+[0.4.0]: https://github.com/ankys/homura_deno/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/ankys/homura_deno/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/ankys/homura_deno/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/ankys/homura_deno/compare/v0.2.0...v0.3.0